requirements.txt
pyaml==19.4.1
PyYAML==5.1.2
